requirements.txt

elevenlabs>=1.0.0
fastapi
firebase-admin
google-genai
langchain_openai
langgraph
moviepy
openai
pandas
pydantic[email]
python-dotenv
uvicorn
websockets>=13.0
litellm
Pillow
python-multipart
google-cloud-firestore>=2.13.0
razorpay
sendgrid>=6.11.0
jinja2>=3.1.3
email-validator>=2.1.0
xhtml2pdf>=0.2.15
